Move home spec to home/index_spec.rb
